Moved mod-operation commands to ModManager. Reordered Utility/Program command dict (grouping).
