+++ OpenRA.Utility/Program.cs
@@ -31,12 +31,13 @@ namespace OpenRA.Utility
 			{ "--map-hash", Command.GetMapHash },
 			{ "--map-preview", Command.GenerateMinimap },
 			{ "--map-upgrade-v5", Command.UpgradeV5Map },
+			{ "--map-import", Command.ImportLegacyMap },
 			{ "--upgrade-map", UpgradeRules.UpgradeMap },
 			{ "--upgrade-mod", UpgradeRules.UpgradeMod },
-			{ "--map-import", Command.ImportLegacyMap },
 			{ "--extract-language-strings", ExtractLanguageStrings.FromMod },
-			{ "--git-install-mod", Command.InstallGitMod },
-			{ "--git-update-mod", Command.UpdateGitMod }
+			{ "--git-mod-install", ModManager.InstallGitMod },
+			{ "--git-mod-upgrade", ModManager.UpdateGitMod },
+			{ "--uninstall-mod", ModManager.DeleteMod }
 		};
 
 		static void Main(string[] args)
